# openapi.yaml
openapi: 3.0.0
info:
  version: "5.0.0-dev"
  title: Metahkg Api
  description: "# Introduction

    The Metahkg Api is a REST api written in fastify/typescript, with MongoDB as the database.
    <br>
    Json is returned on every request.
    <br>
    In this documentation, the base url is /api.

    # Authentication

    Authentication is done with a JWT token, which is sent to the server as a header.
    <br><br>
    ```
    Authorization: Bearer <JWT>
    ```
    <br><br>
    The JWT token is returned on login (/users/login) and verify email (/users/verify).
    <br>
    It is valid for 7 days.
    <br>
    If the JWT token is at least 2 days old, or some of the user info (e.g. user name) has changed, a new token
    is returned via the `token` response header on any requests.
    <br><br>
    ```
    token: <JWT>
    ```
    <br><br>
    Also see [a frontend implmentation using axios](https://gitlab.com/metahkg/metahkg-web/-/blob/dev/src/lib/api.ts).

    # Metahkg api client

    You can also use the [metahkg api client](https://gitlab.com/metahkg/metahkg-api) (in typescript/javascript) to access the metahkg api.
    The client is generated from this specification.
    "
  contact:
    email: wcyat@wcyat.me
  license:
    name: AGPL-3.0-or-later
    url: https://gitlab.com/metahkg/openapi-spec/-/blob/master/LICENSE.md
...
servers:
  - url: https://dev.metahkg.org/api
  - url: http://localhost:3001/api
...
paths:
...
  /threads/{id}/comments/{cid}:
    get:
      tags:
        - comment
      operationId: comment
      security:
        - bearerAuth: []
      summary: Get comment
      description: "Get a comment from a thread with thread id and comment id.
        Authentication required if thread is [hidden](https://docs.metahkg.org/docs/customize/categories/hide).
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/threads/thread/comments/comment.ts)"
      parameters:
        - $ref: "#/components/parameters/ThreadId"
        - $ref: "#/components/parameters/CommentId"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ThreadOrCommentNotFound"
        "410":
          $ref: "#/components/responses/CommentRemoved"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
    patch:
      security:
        - bearerAuth: []
      tags:
        - comment
        - admin
      operationId: commentEdit
      summary: Edit comment
      description: "Edit comment. Admin privilege required.
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/threads/thread/comments/actions/edit.ts)"
      parameters:
        - $ref: "#/components/parameters/ThreadId"
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
                - comment
              properties:
                comment:
                  $ref: "#/components/schemas/Comment/properties/comment"
                reason:
                  description: Reason for editing comment
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ThreadOrCommentNotFound"
        "410":
          $ref: "#/components/responses/CommentRemoved"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
    delete:
      security:
        - bearerAuth: []
      tags:
        - comment
        - admin
      operationId: commentDelete
      summary: Delete comment
      description: "Delete comment. Admin privilege required.
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/threads/thread/comments/actions/delete.ts)"
      parameters:
        - $ref: "#/components/parameters/ThreadId"
        - $ref: "#/components/parameters/CommentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reason
              properties:
                reason:
                  description: Reason for removing comment
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/ThreadOrCommentNotFound"
        "410":
          $ref: "#/components/responses/CommentRemoved"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
...
  /threads/{id}/comments:
    post:
      tags:
        - comment
      operationId: commentCreate
      security:
        - bearerAuth: []
      summary: Create comment
      description: "Create a new comment in the specified thread.
        Authentication is required.
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/threads/thread/comments/create.ts)"
      parameters:
        - $ref: "#/components/parameters/ThreadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
                - rtoken
              properties:
                comment:
                  $ref: "#/components/schemas/Comment/properties/comment"
                rtoken:
                  $ref: "#/components/schemas/rtoken"
                quote:
                  $ref: "#/components/schemas/Comment/properties/id"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                properties:
                  id:
                    $ref: "#/components/schemas/Comment/properties/id"
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden or user muted by an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/ThreadNotFound"
        "429":
          $ref: "#/components/responses/RecaptchaInvalidOrTooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
...
  /categories/{id}:
    get:
      tags:
        - category
      operationId: category
      summary: Get category
      description: "Get a category by id.
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/categories/category.ts)"
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/Invalid"
        "404":
          $ref: "#/components/responses/CategoryNotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
    patch:
      tags:
        - category
        - admin
      operationId: categoryEdit
      summary: Edit a category
      description: "Edit a category. Admin privilege required.
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/categories/actions/edit.ts)"
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  $ref: "#/components/schemas/Category/properties/name"
                tags:
                  $ref: "#/components/schemas/Category/properties/tags"
                pinned:
                  $ref: "#/components/schemas/Category/properties/pinned"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CategoryNotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
    delete:
      tags:
        - category
        - admin
      operationId: categoryDelete
      summary: Delete a category
      description: "Delete a category. Admin privilege required.
        [Source](https://gitlab.com/metahkg/metahkg-server/-/blob/5.0.0-dev/src/routes/categories/actions/delete.ts)"
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "400":
          $ref: "#/components/responses/Invalid"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/CategoryNotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  parameters:
    ThreadId:
      in: path
      name: id
      description: thread id
      required: true
      schema:
        $ref: "#/components/schemas/Thread/properties/id"
    CommentId:
      in: path
      name: cid
      required: true
      description: comment id
      schema:
        $ref: "#/components/schemas/Comment/properties/id"
    CategoryId:
      name: id
      in: path
      description: category id
      required: true
      schema:
        $ref: "#/components/schemas/Category/properties/id"
...
  responses:
    OK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OK"
...
    Invalid:
      description: "Invalid request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: "Unauthorized"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: "Too many requests"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RecaptchaInvalidOrTooManyRequests:
      description: "Recaptcha token invalid, or too many requests"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: "Forbidden"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ThreadNotFound:
      description: "Thread not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ThreadOrCommentNotFound:
      description: "Thread or comment not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    CategoryNotFound:
      description: "Category not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    CommentRemoved:
      description: "Comment removed"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: Bad gateway
      content:
        text/plain:
          schema:
            type: string
            description: "Server error"
        text/html:
          schema:
            type: string
            description: "Server error"
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Emotion:
      type: string
      description: a single emoji
    OK:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          enum:
            - true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: error message
          example: "Error"
...
    User:
      type: object
      required:
        - id
        - name
        - sex
        - role
      properties:
        id:
          type: integer
          minimum: 1
          format: int64
          description: user id
        name:
          type: string
          description: User name
          example: "wcyat"
          minLength: 1
          maxLength: 15
          pattern: ^\S{1,15}$
        sex:
          type: string
          enum:
            - M
            - F
        role:
          type: string
          enum:
            - admin
            - user
    Admin:
      type: object
      description: Admin actions / responses
      properties:
        edits:
          type: array
          items:
            type: object
            required:
              - admin
              - reason
            properties:
              admin:
                $ref: "#/components/schemas/User"
              reason:
                description: reason for editing
                type: string
              date:
                type: string
                format: date-time
        replies:
          description: admin replies
          type: array
          items:
            type: object
            required:
              - admin
              - reply
            properties:
              admin:
                $ref: "#/components/schemas/User"
              reply:
                description: admin's reply
                type: string
              date:
                type: string
                format: date-time
...
    Category:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          minimum: 1
          format: int64
          description: category id
        name:
          type: string
          description: category name
          maxLength: 15
          example: "Chit-chat"
        tags:
          type: array
          description: category tags
          items:
            type: string
            maxLength: 15
        hidden:
          type: boolean
          description: category is hidden
        pinned:
          type: boolean
          description: category is pinned
...
    Comment:
      type: object
      description: Comment object
      required:
        - id
        - user
        - comment
        - text
        - images
        - createdAt
        - slink
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: comment id
        user:
          description: user created the comment
          allOf:
            - $ref: "#/components/schemas/User"
        comment:
          type: string
          description: comment in stringified html
          example: "<p>This is a comment</p>"
        text:
          type: string
          description: comment converted to plain text
          example: "This is a comment"
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image/properties/src"
        createdAt:
          type: string
          format: date-time
        slink:
          type: string
          format: uri
          description: shortened link to the comment
          example: "https://l.metahkg.org/p994Nw2"
        quote:
          allOf:
            - $ref: "#/components/schemas/CommentC"
            - description: The whole comment object of the quoted comment.
              example:
                id: 1
                user:
                  id: 1
                  name: "John Doe"
                  sex: "M"
                  createdAt: "2020-01-01T00:00:00Z"
                comment: "<p>This is a comment</p>"
                text: "This is a comment"
                createdAt: "2020-01-01T00:00:00Z"
                slink: "https://l.metahkg.org/p994Nw2"
        D:
          type: integer
          format: int64
          minimum: 0
          description: number of downvotes
        U:
          type: integer
          format: int64
          minimum: 0
          description: number of upvotes
        replies:
          type: array
          items:
            type: integer
            minimum: 1
            format: int64
          description: list of comment ids that are replies to this comment
        emotions:
          type: array
          description: list of emotions users have expressed
          items:
            type: object
            required:
              - user
              - emotion
            properties:
              user:
                $ref: "#/components/schemas/User/properties/id"
              emotion:
                $ref: "#/components/schemas/Emotion"
        admin:
          $ref: "#/components/schemas/Admin"
    CommentC:
      type: object
      description: "Comment object with constants only (without upvotes, downvotes, replies and emotions)"
      required:
        - id
        - user
        - comment
        - images
        - text
        - createdAt
        - slink
      properties:
        id:
          $ref: "#/components/schemas/Comment/properties/id"
        user:
          $ref: "#/components/schemas/Comment/properties/user"
        comment:
          $ref: "#/components/schemas/Comment/properties/comment"
        text:
          $ref: "#/components/schemas/Comment/properties/text"
        images:
          $ref: "#/components/schemas/Comment/properties/images"
        createdAt:
          $ref: "#/components/schemas/Comment/properties/createdAt"
        slink:
          $ref: "#/components/schemas/Comment/properties/slink"
        quote:
          $ref: "#/components/schemas/Comment/properties/quote"
...
    rtoken:
      type: string
      description: Recaptcha token. See [https://developers.google.com/recaptcha/docs/verify](https://developers.google.com/recaptcha/docs/verify).
      example: "6LeIxAcTAAAAAJcZVRqd76R7tjHpI9u0tRwxV5ksM"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT